seperated sentences
