+++ lib/app/modules/settings/views/settings_page_app_bar.dart
@@ -4,10 +4,10 @@ import 'package:flutter/material.dart';
 
 import 'package:get/get.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:taskwarrior/app/routes/app_pages.dart';
+
 import 'package:taskwarrior/app/utils/constants/taskwarrior_fonts.dart';
 
-import 'package:taskwarrior/app/utils/language/sentences.dart';
+import 'package:taskwarrior/app/utils/language/sentence_manager.dart';
 
 import '../controllers/settings_controller.dart';
