+++ lib/app/utils/language/sentence_manager.dart
@@ -1,0 +1,24 @@
+
+import 'package:taskwarrior/app/utils/language/english_sentences.dart';
+import 'package:taskwarrior/app/utils/language/hindi_sentences.dart';
+import 'package:taskwarrior/app/utils/language/marathi_sentences.dart';
+import 'package:taskwarrior/app/utils/language/sentences.dart';
+import 'package:taskwarrior/app/utils/language/supported_language.dart';
+
+class SentenceManager {
+  final SupportedLanguage currentLanguage;
+
+  SentenceManager({required this.currentLanguage});
+
+  Sentences get sentences {
+    switch (currentLanguage) {
+      case SupportedLanguage.hindi:
+        return HindiSentences();
+      case SupportedLanguage.marathi:
+        return MarathiSentences();
+      case SupportedLanguage.english:
+      default:
+        return EnglishSentences();
+    }
+  }
+}
